tm_event_extractor/src/requirements.txt
functions-framework==3.*
google-cloud-storage==2.6.0
google-cloud-bigquery==3.29.0
requests==2.26.0
pandas==2.2.3
